Implement serach mock
