+++ Makefile
@@ -1,0 +1,5 @@
+
+
+.PHONY: gox
+gox:
+	gox -ldflags=$(LDFLAGS) -output="bin/$(NAME)_{{.OS}}_{{.Arch}}"
