+++ go.mod
@@ -1,3 +1,3 @@
 module github.com/hayashiki/kibela-go
 
-go 1.12
+go 1.13
